// apps/web/src/lib/api/generated/main/model/lootsControllerFetchLootsByGuildIdParams.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */

export type LootsControllerFetchLootsByGuildIdParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit: number;
/**
 * @minimum -9007199254740991
 * @maximum 9007199254740991
 */
cursor: number;
npcs?: string[];
players?: string[];
rarities?: string[];
npcTypes?: string[];
world?: string;
/**
 * @minimum 0
 * @maximum 500
 */
npcLevelMin: number;
/**
 * @minimum 0
 * @maximum 500
 */
npcLevelMax: number;
/**
 * @minimum 0
 * @maximum 500
 */
itemLevelMin: number;
/**
 * @minimum 0
 * @maximum 500
 */
itemLevelMax: number;
/**
 * @minimum 0
 * @maximum 500
 */
playerLevelMin: number;
/**
 * @minimum 0
 * @maximum 500
 */
playerLevelMax: number;
search?: string;
hid?: string;
itemNames?: string[];
/**
 * @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
 */
createdAtMin?: string;
/**
 * @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
 */
createdAtMax?: string;
};